requirements.txt
discord.py==1.3.2
Pillow==7.0.0